+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [1.11.1] - 2022-06-??
+## [1.11.1] - 2022-06-28
+
+### Changed
+
+- Database structure of Workforce table (FK and nullability on "source" column)
+- Remove the "Source" table
 
 ### Fixed
 
 - Set right error message when updating an article abstract
+- workflow/add_workflow resource
 
 ## [1.11.0] - 2022-06-21